5_projects/04_fastapi/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
SQLAlchemy>=2.0.9
aiosqlite>=0.18.0
PyJWT>=2.6.0
bcrypt>=4.0.1
python-multipart>=0.0.6
email-validator>=2.0.0
python-dotenv>=1.0.0
alembic>=1.10.4
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0